+++ tests/testthat/test-htmlwidgets.R
@@ -33,7 +33,7 @@ test_that("New-style widget html methods do not trigger warning on non-tag outpu
 
 test_that("Fallback logic still works", {
   res <- widget_html("does_not_exist", "htmlwidgets", id = "id", style = NULL, class = NULL)
-  expect_identical(res, tags$div(id = "id", "aria-labelledby" = "id-aria"))
+  expect_identical(res, tags$div(id = "id"))
 })
 
 test_that("Legacy methods work with tagList() and HTML()", {
@@ -41,4 +41,11 @@ test_that("Legacy methods work with tagList() and HTML()", {
     widget_html("widgetE", "htmlwidgets", id = "id", style = NULL, class = NULL)
     widget_html("widgetF", "htmlwidgets", id = "id", style = NULL, class = NULL)
   }, NA)
+})
+
+test_that("The widget_html.default respects use_aria option", {
+  res <- widget_html("does_not_exist", "htmlwidgets", id = "id", style = NULL, class = NULL, use_aria = TRUE)
+  expect_identical(res, tags$div(id = "id", "aria-labelledby" = "id-aria"))
+  res <- widget_html("does_not_exist", "htmlwidgets", id = "id", style = NULL, class = NULL, use_aria = FALSE)
+  expect_identical(res, tags$div(id = "id"))
 })